[quant] Reordering the imports in the torch/__init__.py
Because the AO stuff depends on the torch packages, but very few (if any)
torch packages depend on AO, we are moving the imports lower.
That will reduce the probability of cyclic imports, as by the time the
AO would start importing, the rest of the torch would be already imported.

ghstack-source-id: 301c2e0d66f92d66179d38405cc498169a467923
Pull Request resolved: https://github.com/pytorch/pytorch/pull/77065